Split index.ts into smaller submodules
Extracted the types, HTTP utilities, client logic and configuration
into separate files under the ./src folder. The public API itself
remains unchanged.

type: refactor
scope: architecture
