+++ package.json
@@ -44,6 +44,7 @@
   },
   "homepage": "https://github.com/nozzlegear/davenport#readme",
   "devDependencies": {
+    "@eslint/js": "^10.0.1",
     "eslint": "^10.4.1",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-prettier": "^5.5.6",
